requirements.txt
flet==0.5.0
hidapi==0.13.1
python-osc==1.8.1
